Don't manage /etc/toddleglow/config.json as part of the package
Debian 'conf' files will give a warning and prompt if they are part of a
package have been changed a user. ToddleGlow's config is really more of a state
file, so should be changed frequently. This will only copy the default bare
config if there is no config found during the 'configure' phase in the postinst
script

Also restart the service on debian package upgrade